Nouveau répertoire utils Peut effacer des objets dans S3
